+++ src/network_models/copperplate_model.jl
@@ -1,4 +1,4 @@
-function copper_plate(psi_container::PSIContainer, expression::Symbol, bus_count::Int64)
+function copper_plate(psi_container::PSIContainer, expression::Symbol, bus_count::Int)
 
     time_steps = model_time_steps(psi_container)
     devices_netinjection = _remove_undef!(psi_container.expressions[expression])
